move the neo4j-inserter to the kafka-connect
