+++ kafka-connect/neo4j-inserter/insta_posts-inserter/Dockerfile
@@ -4,7 +4,7 @@ COPY go.mod .
 COPY go.sum .
 RUN go mod download
 COPY . .
-RUN CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o neo4j_posts-inserter neo4j-inserter/insta_posts-inserter/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -installsuffix cgo -o neo4j_posts-inserter kafka-connect/neo4j-inserter/insta_posts-inserter/main.go
 
 FROM alpine
 RUN apk --no-cache add ca-certificates
