+++ kafka-connect/neo4j-inserter/insta_user-inserter/main.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"encoding/json"
+	"github.com/codeuniversity/smag-mvp/kafka-connect/neo4j-inserter"
 
 	"github.com/codeuniversity/smag-mvp/kafka"
 	"github.com/codeuniversity/smag-mvp/kafka/changestream"
-	inserter "github.com/codeuniversity/smag-mvp/neo4j-inserter"
 	"github.com/codeuniversity/smag-mvp/service"
 	"github.com/codeuniversity/smag-mvp/utils"
 	bolt "github.com/johnnadratowski/golang-neo4j-bolt-driver"
@@ -15,7 +15,7 @@ func main() {
 	readerConfig := kafka.GetInserterConfig()
 	neo4jConfig := utils.GetNeo4jConfig()
 
-	i := inserter.New(neo4jConfig, kafka.NewReader(readerConfig), insertUsersAndFollowings)
+	i := neo4jinserter.New(neo4jConfig, kafka.NewReader(readerConfig), insertUsersAndFollowings)
 
 	service.CloseOnSignal(i)
 	waitUntilClosed := i.Start()
