+++ kafka-connect/neo4j-inserter/tagged_users-inserter/main.go
@@ -2,20 +2,21 @@ package main
 
 import (
 	"encoding/json"
+	"github.com/codeuniversity/smag-mvp/kafka-connect/neo4j-inserter"
 
 	"github.com/codeuniversity/smag-mvp/kafka"
 	"github.com/codeuniversity/smag-mvp/kafka/changestream"
-	inserter "github.com/codeuniversity/smag-mvp/neo4j-inserter"
 	"github.com/codeuniversity/smag-mvp/service"
 	"github.com/codeuniversity/smag-mvp/utils"
 	bolt "github.com/johnnadratowski/golang-neo4j-bolt-driver"
 )
 
 func main() {
+
 	readerConfig := kafka.GetInserterConfig()
 	neo4jConfig := utils.GetNeo4jConfig()
 
-	i := inserter.New(neo4jConfig, kafka.NewReader(readerConfig), insertPostsAndAddRelationship)
+	i := neo4jinserter.New(neo4jConfig, kafka.NewReader(readerConfig), addTaggedUsersRelationship)
 
 	service.CloseOnSignal(i)
 	waitUntilClosed := i.Start()
@@ -23,25 +24,25 @@ func main() {
 	waitUntilClosed()
 }
 
-type Post struct {
+type taggedUser struct {
 	UserID int `json:"user_id"`
-	PostID int `json:"id"`
+	PostID int `json:"post_	id"`
 }
 
-func insertPostsAndAddRelationship(m *changestream.ChangeMessage, conn bolt.Conn) error {
-	const insertPostsAndAddRelationship = `
+func addTaggedUsersRelationship(m *changestream.ChangeMessage, conn bolt.Conn) error {
+	const addTaggedRelationship = `
 	MERGE(u:USER{id: {userID}})
 	MERGE(p:POST{id: {postID}})
-	MERGE(u)-[:POSTED]->(p)
+	MERGE(u)-[:TAGGED_ON]->(p)
 	`
-	p := &Post{}
-	err := json.Unmarshal(m.Payload.After, p)
+	t := &taggedUser{}
+	err := json.Unmarshal(m.Payload.After, t)
 
 	if err != nil {
 		return err
 	}
 
-	_, err = conn.ExecNeo(insertPostsAndAddRelationship, map[string]interface{}{"userID": p.UserID, "postID": p.PostID})
+	_, err = conn.ExecNeo(addTaggedRelationship, map[string]interface{}{"userID": t.UserID, "postID": t.PostID})
 
 	if err != nil {
 		return err
